once-map: avoid hard-coding `Arc`
The only thing a `OnceMap` really needs to be able to do with the value
is to clone it. Most extant uses benefited from having this done for them
by automatically wrapping values in an `Arc`. But this isn't necessarily
true for all things. For example, a value might have an `Arc` internally
to making cloning cheap in other contexts, and it doesn't make sense to
re-wrap it in an `Arc` just to use it with a `OnceMap`. Or
alternatively, cloning might just be cheap enough on its own that an
`Arc` isn't worth it.

One new recent usage, in `uv-auth`, fell into this category where the
underlying value already had an `Arc` in it. And thus, to get that
underlying value back out, once has to clone the internal `Arc` returned
from `OnceMap`. But with this change, we can just rely on the default
`Clone` impl doing its job.