Parameterize `Requirement` by url type
This change allows switching out the url type for requirements. The original idea was to allow different types for different requirement origins, so that core metadata reads can ban non-pep 508 requirements while we only allow them for requirements.txt. This didn't work out because we expect `&Requirement`s from all sources to match.

I also tried to split pep508_rs into a PEP 508 compliant crate and into our extensions, but they are to tightly coupled.

I think this change is an improvement still as it reduces the hardcoded dependence on `VerbatimUrl`.