Use fs_err for cachedir errors
When running

```
 set UV_CACHE_DIR=%LOCALAPPDATA%\uv\cache-foo && uv venv venv
 ```

 in windows CMD, the error would be just

 ```
 error: The system cannot find the path specified. (os error 3)
 ```

The problem is that the first action in the cache dir is adding the tag, and the `cachedir` crate is using `std::fs` instead of `fs_err`. I've copied the two functions we use from the crate and changed the import from `std::fs` to `fs_err`.

The new error is

```
error: failed to open file `C:\Users\Konstantin\AppData\Local\uv\cache-foo \CACHEDIR.TAG`
  Caused by: The system cannot find the path specified. (os error 3)
```

which correctly explains the problem.

Closes #3280