distribution-types: add FileLocation::to_url
This is useful when one wants a `url::Url` from a `FileLocation`.
I believe we primarily represent `FileLocation` as strings for
performance/serialization reasons, but when generating a lock,
we'd really just like to normalize everything to a URL.

In some cases, a `FileLocation` is generated from a known-valid
`url::Url`, and so that the conversion back to a `url::Url`
*should* be infallible. But this is not, I believe, guaranteed in
all cases.