uv-resolver: add Lock::to_resolution
This commit adds a routine for converting a `Lock` to a
`Resolution`, where a `Resolution` is a map of package names
pinned to a specific version.

I'm not sure that a `Resolution` is ultimately what we want
here (we might need more stuff), but this was the quickest
route I could find to plug a `Lock` into the existing `pip install`
infrastructure.

This commit also does a little refactoring of the `Lock` types.
The main thing is to permit extra state on some of the types
(like a `by_id` map on `Lock` for quick lookups of distributions)
that aren't included in the serialization format of a `Lock`.
We achieve this by defining separate `Wire` types that are
automatically converted to-and-from via `serde`.

Note like with the lock file format types themselves, we leave
a few `todo!()` expressions around. The main idea is to get
something minimally working without spending too much effort
here. (A fair bit of refactoring will be required to generate
a lock file, and it's not clear how much this code will wind up
needing to change anyway.) In particular, we only handle the case
of installing wheels from a registry.