uv: add --unstable-uv-lock-file flag
Like the similar flag on `uv pip compile`, this is meant to be
a temporary toggle to expose "install from lock file" on the CLI.
We expect to remove this eventually.

The larger diff for `pip_install.rs` is mostly just due to moving
some pieces together that we specifically don't want to run (like
resolution) when installing from a lock file.