Fix install_registry_source_dist_cached on Gentoo
Increment the removed file counts in filters
in install_registry_source_dist_cached test, to make it work again
on Gentoo.  The tested counts were updated
in 9a92a3ad3771a78d36214aa958f51bf7bb992f80, but the filters were not.
That said, the respective count increased in Gentoo as well, so adjust
both input and output strings.  I'm updating Windows as a guesswork,
though I suspect that filter may not be necessary anymore, given that
CI was passing.