Use `VerbatimParsedUrl` everywhere
Notice how we remove `?`, error variants and `.unwrap()` in places: The conversion from a `pep508_rs::Requirement` to a `distribution_types::Requirement` is now infallible. Error message improve since we now fail at parse time.

`ParsedPathUrl` feels badly placed, maybe we should move it to `uv-normalize` instead? It's hard to find a good place for a type that's used everywhere but still depends on other crates (uv-git). Or maybe it needs to be the other way round, uv-git should